requirements.txt
#### prod ####
keras
picamera
Pillow
pyaml
pynmea2
pyserial
tensorflow

#### dev ####
matplotlib
jupyter
